Fixes Fix does not take aliases into account #16 (#19)
It's fun when you can fix a bug purely by deleting code... But in this
case we were doing the same job twice - once in code generation and the
second time in post-processing. I've now removed the part that happens
during codegen so that the post-processing can do its job correctly.